get blog post pages loading again
